moved flower related widgets into their own folder
