+++ lib/screens/home/home.dart
@@ -3,7 +3,7 @@ import 'package:redux/redux.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 
 import 'package:learning_compass_exp/store/app_state.dart';
-import 'package:learning_compass_exp/screens/home/widgets/flower.dart';
+import 'package:learning_compass_exp/screens/home/widgets/flower/flower.dart';
 
 import 'package:learning_compass_exp/screens/home/widgets/grid.dart';
