+++ test/widget_tests/flower_widget_test.dart
@@ -4,9 +4,9 @@ import 'package:redux/redux.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'dart:math';
 
-import 'package:learning_compass_exp/screens/home/widgets/flower.dart';
+import 'package:learning_compass_exp/screens/home/widgets/flower/flower.dart';
 import 'package:learning_compass_exp/app.dart';
-import 'package:learning_compass_exp/screens/home/widgets/flower_petal.dart';
+import 'package:learning_compass_exp/screens/home/widgets/flower/flower_petal.dart';
 import 'package:learning_compass_exp/data/models/petal_names.dart';
 import 'package:learning_compass_exp/store/app_state.dart';
 import 'package:learning_compass_exp/store/reducers/app_state_reducer.dart';
